Export default classes in separate statement
Classes that are directly exported with `export default` don't have an encoded class name, which means they can't be [extended](https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation). This change exports the default class as the default in a separate statement so that it can be imported either as the default or as a regular class with an encoded name (that can be extended/augmented).